Improve trace logs in RefCountSearcher
Also lifts the class to top-level
